data-service/requirements.txt
paho-mqtt==2.1.0
bottle==0.13.1
python-dotenv==1.0.0